refactor: upgrade dependencies and overhaul premium access flow
This commit performs a major update to the project's dependency tree, migrates icon usage, and refactors the premium subscription UX from full-screen views to a unified paywall sheet.

Key changes:
- **Dependencies**: Upgraded Flutter SDK to 3.44.0, updated `purchases_flutter`, `syncfusion`, `file_picker`, and `sign_in_with_apple`.
- **Premium UX**: Replaced multiple `PremiumScreen` implementations with a centralized `showPremiumPaywall()` sheet for a more seamless user experience.
- **Icons**: Migrated from `lucide_icons` to `lucide_icons_flutter`.
- **Navigation**: Updated default routing to land on `marketPlace` and refactored `BottomNavigationBar` indices and labels.
- **AI Features**: Enhanced `AiPromoteController` to support Product and Service post types, including price and detail context in AI prompts.
- **Marketplace**: Added location-based filtering and search capabilities to `MarketController` and `BuyerRequestController`.
- **Performance**: Implemented caching for `rankWinner` in `HomeController` to reduce API calls and UI jitter.
- **Infrastructure**: Updated Android Gradle configurations and added `.env.example`.

+++ lib/bbpro/models/product_model.dart
@@ -79,8 +79,9 @@ class Product {
       isActive: json['isActive'],
       storageLocation: json['storageLocation'],
       productNumber: json['productNumber'].toString(),
-      quantity:
-          json['quantity'] != null ? int.parse(json['quantity'].toString()) : 0,
+      quantity: json['quantity'] != null
+          ? double.parse(json['quantity'].toString()).toInt()
+          : 0,
       startAt: json['startAt'] == null ? null : DateTime.parse(json['startAt']),
       endAt: json['endAt'] == null ? null : DateTime.parse(json['endAt']),
       color: json['color'] != null ? List<String>.from(json['color']) : null,
